capitalOne/requirements.txt
Django==1.8.5
pinax-theme-bootstrap==7.1.1
django-user-accounts==1.2
metron==1.3.5
pinax-eventlog==1.0.0
django-jsonfield==0.9.15
django-crispy-forms==1.5.2
requests==2.7.0
pandas==0.17.0
wolframalpha==1.2
geopy==1.11.0

